pyproject.toml
[tool.poetry]
name = "historical-quest"
version = "1.0.0"
description = "Telegram bot with historical quests"
authors = ["Maksim Gusev <51376126+AlanTheKnight@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "historical_quest" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.8"
djangorestframework = "^3.14.0"
django-jsonform = "^2.21.2"
aiogram = "^3.2.0"
toml = "^0.10.2"
requests = "^2.31.0"
pillow = "^10.1.0"
django-filter = "^23.4"
django-imagekit = "^5.0.0"
requests-cache = "^1.1.1"
django-cleanup = "^8.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
